Day 1 complete!
This was a very approachable problem overall. Heavy recursion use here! I
haven't gotten to use that at work much, so it was very exciting to walk through
everything. For now I commented out the debug statements that helped me walk
through everything, and I was able to add some logging at the last minute. That
was pretty cool if I do say so myself.